Revert "This is a sloppy commit. I have done too much work between this one and the last."
This reverts commit bcaa58cce60eca608c879b1febeea39c3ff8717a.

+++ exp/include/env/constants.h
@@ -22,15 +22,15 @@
 #include "utility/io.h"
 
 typedef struct Constants {
-    u32 count;
-    u32 capacity;
+    u64 count;
+    u64 capacity;
     Value *buffer;
 } Constants;
 
 void constants_initialize(Constants *constants);
 void constants_terminate(Constants *constants);
 Operand constants_append(Constants *constants, Value value);
-Value *constants_at(Constants *constants, u32 index);
+Value *constants_at(Constants *constants, u16 index);
 
 struct Context;
 void print_constants(Constants const *constants,
